correction in cookbook listing api
