+++ src/workstation-gui/app/helpers/application_helper.rb
@@ -1,0 +1,3 @@
+module ApplicationHelper
+
+end
